requirements.txt
flask
python-dotenv
flask-smorest
flask-sqlalchemy